Realm worlds: readable geography, fog of war, and a pathing fix
Realm maps are generated on their own now: forest belts, rock ridges
broken by passes, one or two rivers with fords, cleared town sites, a
mine per region, and a carved route so every region can be reached from
the capital. Large worlds are 90 tiles wide. The camera opens on the
home village and can zoom out to half scale on big maps.

Fog of war: the player explores by moving. Unexplored land is black,
explored land is dimmed, enemies show only in sight. Scouts see
farthest. The explored map is saved run-length encoded.

The dijkstra behind every flow field stored float32 distances but
compared popped keys as doubles, so any node reached through a diagonal
step was dropped as stale. Paths were quietly truncated. Fixed by
pushing the rounded value. Better paths made timing pushes deadlier, so
the AI now saves for stronger units against a budget of twenty seconds
of income, sallies only against a force it outmatches, and engages
anything already at its gates. Garrison need comes from hostile
neighbors only.

+++ src/sim/pathing.ts
@@ -47,7 +47,8 @@ export function dijk(cols: number, rows: number, starts: number | number[], cost
         const ni = ny * cols + nx, c = costFn(ni);
         if (c === Infinity) continue;
         if (dx && dy && (costFn(cy * cols + nx) === Infinity || costFn(ny * cols + cx) === Infinity)) continue;
-        const nd = dist[cur] + c * (dx && dy ? 1.414 : 1);
+        // dist is float32: push the value it will actually hold, or the pop-time staleness check drops the node.
+        const nd = Math.fround(dist[cur] + c * (dx && dy ? 1.414 : 1));
         if (nd < dist[ni]) { dist[ni] = nd; push(nd, ni); }
       }
   }
